+++ src/main/java/com/example/demo/VO/Title.java
@@ -22,7 +22,7 @@ public class Title {
     private BigInteger paragraphId;
 
     @ApiModelProperty("行距")
-    private Integer lineSpacing;
+    private Double lineSpacing;
 
     @ApiModelProperty("左方缩排")
     private Integer indentFromLeft;
@@ -52,11 +52,11 @@ public class Title {
         this.paragraphId = paragraphId;
     }
 
-    public Integer getLineSpacing() {
+    public Double getLineSpacing() {
         return lineSpacing;
     }
 
-    public void setLineSpacing(Integer lineSpacing) {
+    public void setLineSpacing(Double lineSpacing) {
         this.lineSpacing = lineSpacing;
     }
